test: add t.Run, fix parallel-unsafe chdir, rename shell tests, move shared helper
